Admin endpoint base structure
