+++ angular-spa-ui/js/admin-app/admin-app-srv.js
@@ -1,0 +1,5 @@
+"use strict";
+
+module.exports = function (app) {
+
+}
