+++ angular-spa-ui/js/common/app-config.js
