+++ angular-spa-ui/js/common/common-ctrl.js
@@ -1,9 +1,9 @@
 "use strict";
 
 module.exports = function (app) {
-
+    
     app.controller('mainCtrl', function ($scope, promises, getTemplate, appConfig) {
-
+        
         var config = appConfig.config;
         var paths = config.paths;
         var methods = config.methods;
@@ -71,4 +71,5 @@ module.exports = function (app) {
                     console.log('Error '+ error.status);
                 });
     });
+    
 }
