+++ angular-spa-ui/js/common/index.js
@@ -1,1 +1,10 @@
 "use strict";
+
+module.exports = function (app) {
+
+    require('./common-srv')(app);
+    
+    // Page preload actions realized in mainCtrl
+    require('./common-ctrl')(app);
+
+};
