--- angular-spa-ui/js/search-app/bootstrap.js
@@ -1,8 +1,0 @@
-"use strict";
-
-module.exports = function (app) {
-
-    require('./search-app-srv')(app);
-    require('./search-app-ctrl')(app);
-
-};
